api/requirements.txt
Flask>=2.0.2
psycopg2-binary>=2.8.4
Flask-SQLAlchemy>=2.4.1
Flask-Migrate>=2.5.2
bcrypt
PyJWT